*[development]

+++ MT.MicroService.Services.Person/Controllers/ContactInfoController.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using MT.MicroService.Core.Entity;
 using MT.MicroService.Core.Services;
 using MT.MicroService.Services.Person.Dtos;
 using System;
@@ -25,20 +26,25 @@ namespace MT.MicroService.Services.Person.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> KisiyeIletisimBilgisiEkle(ContactInfoDto contactInfoDto)
+        public async Task<IActionResult> AddContactInfotoPerson(ContactInfoDto contactInfoDto)
         {
             var newContactInfo = await _contactInfoService.AddAsync(_mapper.Map<MT.MicroService.Core.Entity.ContactInfo>(contactInfoDto));
             return Created(string.Empty, _mapper.Map<ContactInfoDto>(newContactInfo));
+            
 
         }
 
-        [HttpDelete("{id}")]
-        public IActionResult KisidenIletisimBilgisiKaldirma(int id)
+        [HttpDelete]
+        public IActionResult RemoveContactInfoFromPerson(int id)
         {
             var contactInfo = _contactInfoService.GetByIdAsync(id).Result;
-            _contactInfoService.Delete(contactInfo);
-            return NoContent();
+            if (contactInfo != null)
+            {
+                _contactInfoService.Delete(contactInfo);
 
+                return NoContent();
+            }
+            return BadRequest("Silinecek kayıt yok");
         }
     }
 }
